+++ Delivery 1 - Base Version/Shared/Interfaces/IServerClientCommands.cs
