+++ Delivery 1 - Base Version/client/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace client
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World Client!");
+        }
+    }
+}
