+++ Delivery 1 - Base Version/dida-gstore.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DIDA-GSTORE", "dida-gstore.csproj", "{62D267E2-E993-43A2-934E-71DCBCE7D510}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
