--- Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace DIDA_GSTORE
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
