--- client/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace client
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World Client!");
-        }
-    }
-}
