--- puppetMaster/puppetMaster.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\processCreationService\processCreationService.csproj">
-      <Name>processCreationService</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\server\server.csproj">
-      <Name>server</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Shared\Shared.csproj">
-      <Name>Shared</Name>
-    </ProjectReference>
-  </ItemGroup>
-</Project>
