--- server/Program.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using Shared;
-
-namespace server
-{
-    class Program : IServerClientCommands
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World Server!");
-        }
-
-        public void getStatus()
-        {
-            Console.WriteLine(">>> Status");
-        }
-    }
-}
